閲覧部分をクライアントコンポーネントに切り出し
